refactor plugging-in of RuntimeCommand, move 3 workspace commands to runtime
