+++ dlt/_workspace/cli/commands.py
@@ -829,49 +829,11 @@ workspace info.
             default=None,
         )
 
-        subparsers.add_parser(
-            "login",
-            help=(
-                "Login to dltHub Runtime using Github OAuth and connect current workspace to the"
-                " remote one"
-            ),
-            description="Login to dltHub Runtime using Github OAuth",
-        )
-
-        subparsers.add_parser(
-            "logout",
-            help="Logout from dltHub Runtime",
-            description="Logout from dltHub Runtime",
-        )
-
-        deploy_cmd = subparsers.add_parser(
-            "deploy",
-            help="Create, run and inspect scripts in runtime",
-            description="Manipulate scripts in workspace",
-        )
-        deploy_cmd.add_argument("script_name", help="Local path to the script")
-        deploy_cmd.add_argument(
-            "--profile",
-            "-p",
-            nargs="?",
-            help="Profile to use for the run",
-        )
-        deploy_cmd.add_argument(
-            "-i",
-            "--interactive",
-            action="store_true",
-            help=(
-                "Whether the script should be deployed as interactive (e.g. a notebook). False by"
-                " default"
-            ),
-        )
-
     def execute(self, args: argparse.Namespace) -> None:
         from dlt._workspace._workspace_context import active
-        from dlt._workspace.cli._runtime_command import deploy, login, logout
         from dlt._workspace.cli._workspace_command import (
-            print_workspace_info,
             clean_workspace,
+            print_workspace_info,
             show_workspace,
             start_mcp,
         )
@@ -886,12 +848,6 @@ workspace info.
             show_workspace(workspace_context, args.edit)
         elif args.workspace_command == "mcp":
             start_mcp(workspace_context, port=args.port, stdio=args.stdio)
-        elif args.workspace_command == "login":
-            login(minimal_logging=False)
-        elif args.workspace_command == "logout":
-            logout()
-        elif args.workspace_command == "deploy":
-            deploy(args.script_name, bool(args.interactive), args.profile)
         else:
             self.parser.print_usage()
 
@@ -952,3 +908,325 @@ Pin a profile to the Workspace, this will be the new default profile while it is
             pin_profile(workspace_context, args.profile_name)
         else:
             self.parser.print_usage()
+
+
+class RuntimeCommand(SupportsCliCommand):
+    command = "runtime"
+    help_string = "Connect to dltHub Runtime and run your code remotely"
+    description = """
+    Allows to connect to the dltHub Runtime, deploy and run local workspaces there. Requires dltHub license.
+    """
+
+    def configure_parser(self, parser: argparse.ArgumentParser) -> None:
+        self.parser = parser
+
+        subparsers = parser.add_subparsers(
+            title="Available subcommands", dest="runtime_command", required=False
+        )
+
+        subparsers.add_parser(
+            "login",
+            help=(
+                "Login to dltHub Runtime using Github OAuth and connect current workspace to the"
+                " remote one"
+            ),
+            description="Login to dltHub Runtime using Github OAuth",
+        )
+
+        subparsers.add_parser(
+            "logout",
+            help="Logout from dltHub Runtime",
+            description="Logout from dltHub Runtime",
+        )
+
+        deploy_cmd = subparsers.add_parser(
+            "deploy",
+            help="Create, run and inspect scripts in runtime",
+            description="Manipulate scripts in workspace",
+        )
+        self._configure_deploy_parser(deploy_cmd)
+
+        # run a script
+        run_cmd = subparsers.add_parser(
+            "run",
+            help="Run a script in the Runtime",
+            description="Create or update a script and trigger a run",
+        )
+        self._configure_run_parser(run_cmd)
+
+        # runs management
+        runs_cmd = subparsers.add_parser(
+            "runs",
+            help="Manipulate runs in workspace",
+            description="Manipulate runs in workspace",
+        )
+        self._configure_runs_parser(runs_cmd)
+
+        # deployments
+        deployment_cmd = subparsers.add_parser(
+            "deployment",
+            help="Manipulate deployments in workspace",
+            description="Manipulate deployments in workspace",
+        )
+        self._configure_deployments_parser(deployment_cmd)
+
+        # scripts
+        script_cmd = subparsers.add_parser(
+            "script",
+            help="Create, list and inspect scripts in runtime",
+            description="Manipulate scripts in workspace",
+        )
+        self._configure_scripts_parser(script_cmd)
+
+        # configurations
+        configuration_cmd = subparsers.add_parser(
+            "configuration",
+            help="Manipulate configurations in workspace",
+            description="Manipulate configurations in workspace",
+        )
+        self._configure_configurations_parser(configuration_cmd)
+
+    def _configure_deploy_parser(self, deploy_cmd: argparse.ArgumentParser) -> None:
+        deploy_cmd.add_argument("script_name", help="Local path to the script")
+        deploy_cmd.add_argument(
+            "--profile",
+            "-p",
+            nargs="?",
+            help="Profile to use for the run",
+        )
+        deploy_cmd.add_argument(
+            "-i",
+            "--interactive",
+            action="store_true",
+            help=(
+                "Whether the script should be deployed as interactive (e.g. a notebook). False by"
+                " default"
+            ),
+        )
+
+    def _configure_runs_parser(self, runs_cmd: argparse.ArgumentParser) -> None:
+        runs_cmd.add_argument(
+            "script_name_or_run_id",
+            nargs="?",
+            help="The name of the script we're working with or the id of the run of this script",
+        )
+        runs_cmd.add_argument(
+            "--list", "-l", action=argparse.BooleanOptionalAction, help="List all runs in workspace"
+        )
+        runs_subparsers = runs_cmd.add_subparsers(
+            title="Available subcommands", dest="operation", required=False
+        )
+
+        # manage runs
+        runs_subparsers.add_parser(
+            "list",
+            help="List all runs of the script, only works if script name is provided",
+            description="List all runs of the script, only works if script name is provided",
+        )
+        runs_subparsers.add_parser(
+            "info",
+            help="Get detailed information about a run",
+            description="Get detailed information about a run",
+        )
+        runs_subparsers.add_parser(
+            "logs",
+            help="Get the logs of a run",
+            description="Get the logs of a run",
+        )
+        runs_subparsers.add_parser(
+            "cancel",
+            help="Cancel a run in the Runtime",
+            description="Cancel a run in the Runtime",
+        )
+
+    def _configure_run_parser(self, run_cmd: argparse.ArgumentParser) -> None:
+        run_cmd.add_argument(
+            "script_name_or_id",
+            help="Local path to the script or id/name of deployed script",
+        )
+        run_cmd.add_argument(
+            "--profile",
+            "-p",
+            nargs="?",
+            help="Profile to use for the run",
+        )
+        run_cmd.add_argument(
+            "-i",
+            "--interactive",
+            action="store_true",
+            help=(
+                "Whether the script should be deployed as interactive (e.g. a notebook). False by"
+                " default"
+            ),
+        )
+
+    def _configure_deployments_parser(self, deployment_cmd: argparse.ArgumentParser) -> None:
+        deployment_cmd.add_argument(
+            "--list",
+            "-l",
+            action=argparse.BooleanOptionalAction,
+            help="List all deployments in workspace",
+        )
+        deployment_cmd.add_argument("deployment_id", nargs="?", help="Deployment id (UUID)")
+        deployment_subparsers = deployment_cmd.add_subparsers(
+            title="Available subcommands", dest="operation", required=False
+        )
+        deployment_subparsers.add_parser(
+            "info",
+            help="Get detailed information about a deployment",
+            description="Get detailed information about a deployment",
+        )
+        deployment_subparsers.add_parser(
+            "sync",
+            help="Create new deployment if local workspace content changed",
+            description="Create new deployment if local workspace content changed",
+        )
+
+    def _configure_scripts_parser(self, script_cmd: argparse.ArgumentParser) -> None:
+        script_cmd.add_argument(
+            "script_name_or_id",
+            nargs="?",
+            help="Local path to the script or id/name of deployed script",
+        )
+        script_cmd.add_argument(
+            "--list",
+            "-l",
+            action=argparse.BooleanOptionalAction,
+            help="List all scripts in workspace",
+        )
+        script_subparsers = script_cmd.add_subparsers(
+            title="Available subcommands", dest="operation", required=False
+        )
+        script_subparsers.add_parser(
+            "info",
+            help="Get detailed information about a script",
+            description="Get detailed information about a script",
+        )
+        script_subparsers.add_parser(
+            "sync",
+            help="Create or update the script",
+            description="Create or update the script",
+        )
+
+    def _configure_configurations_parser(self, configuration_cmd: argparse.ArgumentParser) -> None:
+        configuration_cmd.add_argument(
+            "--list",
+            "-l",
+            action=argparse.BooleanOptionalAction,
+            help="List all configurations in workspace",
+        )
+        configuration_cmd.add_argument(
+            "configuration_id", nargs="?", help="Configuration id (UUID)"
+        )
+        configuration_subparsers = configuration_cmd.add_subparsers(
+            title="Available subcommands", dest="operation", required=False
+        )
+        configuration_subparsers.add_parser(
+            "info",
+            help="Get detailed information about a configuration",
+            description="Get detailed information about a configuration",
+        )
+        configuration_subparsers.add_parser(
+            "sync",
+            help="Create new configuration if local config content changed",
+            description="Create new configuration if local config content changed",
+        )
+
+    def execute(self, args: argparse.Namespace) -> None:
+        from dlt._workspace.cli._runtime_command import (
+            deploy,
+            fetch_run_logs,
+            get_api_client,
+            get_configuration_info,
+            get_configurations,
+            get_deployment_info,
+            get_deployments,
+            get_run_info,
+            get_runs,
+            get_script_info,
+            get_scripts,
+            login,
+            logout,
+            request_run_cancel,
+            run_script,
+            sync_configuration,
+            sync_deployment,
+        )
+
+        if args.runtime_command == "login":
+            login(minimal_logging=False)
+        elif args.runtime_command == "logout":
+            logout()
+        elif args.runtime_command == "deploy":
+            deploy(args.script_name, bool(args.interactive), args.profile)
+        else:
+            auth_service = login()
+            api_client = get_api_client(auth_service)
+            if args.runtime_command == "run":
+                run_script(
+                    args.script_name_or_id,
+                    bool(args.interactive),
+                    args.profile,
+                    auth_service=auth_service,
+                    api_client=api_client,
+                )
+            elif args.runtime_command == "runs":
+                if args.list:
+                    get_runs(auth_service=auth_service, api_client=api_client)
+                elif args.operation == "info":
+                    get_run_info(
+                        script_name_or_run_id=args.script_name_or_run_id,
+                        auth_service=auth_service,
+                        api_client=api_client,
+                    )
+                elif args.operation == "logs":
+                    fetch_run_logs(
+                        script_name_or_run_id=args.script_name_or_run_id,
+                        auth_service=auth_service,
+                        api_client=api_client,
+                    )
+                elif args.operation == "cancel":
+                    request_run_cancel(
+                        script_name_or_run_id=args.script_name_or_run_id,
+                        auth_service=auth_service,
+                        api_client=api_client,
+                    )
+                elif args.operation == "list":
+                    get_runs(
+                        script_name=args.script_name_or_run_id,  # raising downstream if run_id is passed
+                        auth_service=auth_service,
+                        api_client=api_client,
+                    )
+            elif args.runtime_command == "deployment":
+                if args.list:
+                    get_deployments(auth_service=auth_service, api_client=api_client)
+                elif args.operation == "info":
+                    get_deployment_info(
+                        deployment_id=args.deployment_id,
+                        auth_service=auth_service,
+                        api_client=api_client,
+                    )
+                elif args.operation == "sync":
+                    sync_deployment(auth_service=auth_service, api_client=api_client)
+            elif args.runtime_command == "script":
+                if args.list:
+                    get_scripts(auth_service=auth_service, api_client=api_client)
+                elif args.operation == "info":
+                    get_script_info(
+                        script_id_or_name=args.script_name_or_id,
+                        auth_service=auth_service,
+                        api_client=api_client,
+                    )
+            elif args.runtime_command == "configuration":
+                if args.list:
+                    get_configurations(auth_service=auth_service, api_client=api_client)
+                elif args.operation == "info":
+                    get_configuration_info(
+                        configuration_id=args.configuration_id,
+                        auth_service=auth_service,
+                        api_client=api_client,
+                    )
+                elif args.operation == "sync":
+                    sync_configuration(auth_service=auth_service, api_client=api_client)
+            else:
+                self.parser.print_usage()
